Adding Golang support
A couple of notes:

* The debian Go package builds are frozen and no longer up to date
[[source](https://launchpad.net/~gophers/+archive/go)]
* Installing Go with `gvm` instead of building from source was
preferred because it is faster and much easier to manage different
versions of Go
* `gvm` is being installed as the root user because it is not necessary
for the app user to be able to compile the Go application in order to
run it. It is recommended that the application be built as root and then
moved to the correct location with the correct permissions so that it
may be run by the app user.